chore:delete redundant code.
chore:delete redundant code.
