+++ sample/src/main/AndroidManifest.xml
@@ -9,10 +9,10 @@
         android:name=".SampleApplication"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
+        android:networkSecurityConfig="@xml/network_security_config"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        tools:ignore="AllowBackup,GoogleAppIndexingWarning"
-        android:networkSecurityConfig="@xml/network_security_config">
+        tools:ignore="AllowBackup,GoogleAppIndexingWarning">
 
         <meta-data
             android:name="com.airwallex.sample.metadata.api_key"
@@ -52,11 +52,16 @@
         </activity>
 
         <activity
-            android:name="com.airwallex.paymentacceptance.ui.PaymentCartActivity"
-            android:exported="true"
-            android:launchMode="singleTask"
+            android:name="com.airwallex.paymentacceptance.ui.MainActivity"
             android:screenOrientation="portrait"
-            android:theme="@style/AirwallexDefaultTheme"
+            android:theme="@style/ActivityAnimTheme"
+            tools:ignore="LockedOrientationActivity" />
+
+        <activity
+            android:name="com.airwallex.paymentacceptance.ui.UIIntegrationActivity"
+            android:exported="true"
+            android:screenOrientation="portrait"
+            android:theme="@style/ActivityAnimTheme"
             tools:ignore="LockedOrientationActivity">
 
             <intent-filter>
@@ -70,19 +75,8 @@
                     android:host="${applicationId}"
                     android:scheme="airwallexcheckout" />
             </intent-filter>
+
         </activity>
-
-        <activity
-            android:name="com.airwallex.paymentacceptance.ui.MainActivity"
-            android:screenOrientation="portrait"
-            android:theme="@style/ActivityAnimTheme"
-            tools:ignore="LockedOrientationActivity" />
-
-        <activity
-            android:name="com.airwallex.paymentacceptance.ui.UIIntegrationActivity"
-            android:screenOrientation="portrait"
-            android:theme="@style/ActivityAnimTheme"
-            tools:ignore="LockedOrientationActivity" />
 
         <activity
             android:name="com.airwallex.paymentacceptance.ui.APIIntegrationActivity"
