+++ sample/src/main/res/drawable/ic_menu.xml
@@ -5,6 +5,6 @@
     android:viewportHeight="20">
   <path
       android:pathData="M4.6483,17.028L5.5032,17.0508C5.7185,17.0566 5.9163,17.1687 6.029,17.3489L6.4766,18.0644C6.8166,18.608 7.2824,19.0652 7.8361,19.3991C9.782,20.5725 12.3283,19.9749 13.5234,18.0644L13.971,17.3489C14.0837,17.1687 14.2815,17.0566 14.4968,17.0508L15.3517,17.028C16.0012,17.0107 16.6374,16.8433 17.2087,16.5395C19.2167,15.4716 19.9627,13.0078 18.8751,11.0364L18.4678,10.298C18.3652,10.1121 18.3652,9.8879 18.4678,9.702L18.8751,8.9636C19.1846,8.4027 19.355,7.7781 19.3727,7.1404C19.4347,4.8991 17.6345,3.0329 15.3517,2.972L14.4968,2.9492C14.2815,2.9434 14.0837,2.8313 13.971,2.6511L13.5234,1.9356C13.1834,1.392 12.7176,0.9348 12.164,0.6009C10.218,-0.5725 7.6717,0.0251 6.4766,1.9356L6.029,2.6511C5.9163,2.8313 5.7185,2.9434 5.5032,2.9492L4.6483,2.972C3.9988,2.9893 3.3626,3.1567 2.7913,3.4605C0.7833,4.5284 0.0373,6.9922 1.1249,8.9636L1.5323,9.702C1.6348,9.8879 1.6348,10.1121 1.5323,10.298L1.1249,11.0364C0.8154,11.5973 0.645,12.2219 0.6273,12.8596C0.5653,15.1009 2.3655,16.9671 4.6483,17.028ZM8.2015,16.0389C7.6379,15.138 6.649,14.5774 5.5725,14.5487L4.7176,14.5259C3.8423,14.5025 3.152,13.787 3.1758,12.9276C3.1826,12.6831 3.2479,12.4436 3.3666,12.2285L3.7739,11.4902C4.2868,10.5605 4.2868,9.4395 3.7739,8.5098L3.3666,7.7715C2.9496,7.0156 3.2356,6.0709 4.0055,5.6614C4.2246,5.5449 4.4685,5.4807 4.7176,5.4741L5.5725,5.4513C6.649,5.4226 7.6379,4.862 8.2015,3.9611L8.649,3.2456C9.1073,2.513 10.0836,2.2839 10.8297,2.7338C11.042,2.8618 11.2206,3.0371 11.351,3.2456L11.7986,3.9611C12.3621,4.862 13.351,5.4226 14.4275,5.4513L15.2824,5.4741C16.1577,5.4974 16.848,6.213 16.8242,7.0724C16.8174,7.3169 16.7521,7.5564 16.6334,7.7715L16.2261,8.5098C15.7132,9.4395 15.7132,10.5605 16.2261,11.4902L16.6334,12.2285C17.0504,12.9844 16.7644,13.9291 15.9945,14.3386C15.7754,14.4551 15.5315,14.5193 15.2824,14.5259L14.4275,14.5487C13.351,14.5774 12.3621,15.138 11.7986,16.0389L11.351,16.7544C10.8927,17.487 9.9164,17.7161 9.1703,17.2662C8.958,17.1382 8.7794,16.9629 8.649,16.7544L8.2015,16.0389ZM10,13.75C7.9289,13.75 6.25,12.0711 6.25,10C6.25,7.9289 7.9289,6.25 10,6.25C12.0711,6.25 13.75,7.9289 13.75,10C13.75,12.0711 12.0711,13.75 10,13.75ZM11.25,10C11.25,10.6904 10.6904,11.25 10,11.25C9.3096,11.25 8.75,10.6904 8.75,10C8.75,9.3096 9.3096,8.75 10,8.75C10.6904,8.75 11.25,9.3096 11.25,10Z"
-      android:fillColor="#FFFFFF"
+      android:fillColor="@color/color_primary"
       android:fillType="evenOdd"/>
 </vector>
