--- sample/src/main/res/layout/fragment_cart.xml
@@ -1,142 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <ScrollView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_above="@id/btnCheckout"
-        android:paddingBottom="16dp">
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical">
-
-            <TextView
-                style="@style/AirwallexPaymentHeader"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:paddingLeft="24dp"
-                android:paddingRight="24dp"
-                android:text="@string/my_cart" />
-
-            <TextView
-                style="@style/AirwallexPaymentSubHeaderBold"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="24dp"
-                android:paddingLeft="24dp"
-                android:paddingRight="24dp"
-                android:text="@string/shipping" />
-
-            <com.airwallex.paymentacceptance.ui.widget.ShippingItemView
-                android:id="@+id/shippingItemView"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:paddingHorizontal="24dp" />
-
-            <TextView
-                style="@style/AirwallexPaymentSubHeaderBold"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="24dp"
-                android:paddingLeft="24dp"
-                android:paddingRight="24dp"
-                android:text="@string/item" />
-
-            <LinearLayout
-                android:id="@+id/llProducts"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:background="@android:color/white"
-                android:orientation="vertical"
-                android:paddingLeft="24dp"
-                android:paddingRight="24dp" />
-
-            <RelativeLayout
-                android:layout_width="match_parent"
-                android:layout_height="58dp"
-                android:background="@android:color/white"
-                android:orientation="vertical"
-                android:paddingLeft="24dp"
-                android:paddingRight="24dp">
-
-                <TextView
-                    android:id="@+id/tvSubtotalLabel"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_centerVertical="true"
-                    android:text="@string/subtotal"
-                    android:textColor="@color/airwallex_color_grey_70"
-                    android:textSize="15sp" />
-
-                <TextView
-                    android:id="@+id/tvOrderSubtotalPrice"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_alignParentEnd="true"
-                    android:layout_centerVertical="true"
-                    android:layout_toEndOf="@id/tvSubtotalLabel"
-                    android:gravity="end"
-                    android:textColor="@color/airwallex_color_grey_100"
-                    android:textSize="15sp"
-                    android:textStyle="bold" />
-
-            </RelativeLayout>
-
-            <View
-                android:layout_width="match_parent"
-                android:layout_height="1dp"
-                android:layout_marginHorizontal="24dp"
-                android:background="@drawable/airwallex_line_divider" />
-
-            <RelativeLayout
-                android:layout_width="wrap_content"
-                android:layout_height="58dp"
-                android:background="@android:color/white"
-                android:paddingLeft="24dp"
-                android:paddingRight="24dp">
-
-                <TextView
-                    android:id="@+id/tvTotalLabel"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_centerVertical="true"
-                    android:text="@string/total"
-                    android:textColor="@color/airwallex_color_grey_70"
-                    android:textSize="15sp" />
-
-                <TextView
-                    android:id="@+id/tvOrderTotalPrice"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_alignParentEnd="true"
-                    android:layout_centerVertical="true"
-                    android:layout_toEndOf="@id/tvTotalLabel"
-                    android:gravity="end"
-                    android:textColor="@color/airwallex_color_grey_100"
-                    android:textSize="15sp"
-                    android:textStyle="bold" />
-
-            </RelativeLayout>
-        </LinearLayout>
-    </ScrollView>
-
-    <androidx.appcompat.widget.AppCompatButton
-        android:id="@+id/btnCheckout"
-        android:layout_width="match_parent"
-        android:layout_height="44dp"
-        android:layout_alignParentBottom="true"
-        android:layout_margin="24dp"
-        android:background="@drawable/airwallex_button_solid"
-        android:gravity="center"
-        android:padding="0dp"
-        android:text="@string/checkout"
-        android:textColor="@android:color/white"
-        android:textSize="14sp" />
-
-</RelativeLayout>
