--- sample/src/main/res/menu/menu_cart.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <item
-        android:icon="@drawable/ic_menu"
-        android:title=""
-        app:showAsAction="ifRoom">
-
-        <menu>
-            <item
-                android:id="@+id/weChat_demo"
-                android:title="@string/wechat_demo"
-                app:showAsAction="never" />
-            <item
-                android:id="@+id/h5demo"
-                android:title="@string/h5demo"
-                app:showAsAction="never" />
-            <item
-                android:id="@+id/settings"
-                android:title="@string/settings"
-                app:showAsAction="never" />
-
-        </menu>
-    </item>
-</menu>
